Fixing #345 - where outdated conversation id could not be found in smarti
